Create simple_werewolf.py
I tried to learn to use dictionary, list, random library in this program.
In this program, you're going to play a simple werewolf game with my 5 instructors from CIP. There is only 1 werewolf, and the rest will be villagers. You're always the chief of this village. To save your village, find the werewolf among villagers. You'll lose if there is 1 name left (which means the werewolf is going to eat you)