Create madlibs.py
Task here:
Mad Libs is a word game where players are prompted for one word at a time, and the words are eventually filled into the blanks of a word template to make an entertaining story! In the starter code we've provided most of a short story, with constants WIZARD, NUMBER_OF, FRUIT, PRICE, and YEARS interspersed throughout! 

Start by taking a close look at the starter code and running it as is to see how the story goes so far. Notice that if you change the value of, say, the WIZARD constant at the top and rerun the program, it'll change throughout the story! 

Your job is to finish writing the last line of this story, which should go like this (recall that YEARS = 300 and WIZARD = 'Karel the Omniscient'; you should use both these constants in your line of code)