Create subtract_numbers.py
Task here:
Write a program in the file subtract_numbers.py that reads two real numbers from the user and prints the first number minus the second number.

You can assume the user will always enter valid real numbers as input (negative values are fine). Yes, we know this problem is really similar to a problem we did in class – that’s why this problem is a sandcastle!