Create 8ball.py
Task here:

The idea behind an 8-ball is very simple. You ask it a yes or no question, and it tells you the answer. The catch is that the answer it chooses is randomly selected from a set of prefabricated responses.

Your job is to write a program that continuously prompts the user for a yes or no question, and then randomly selects from a set of canned answers. The classic 8-ball responses were:

As I see it, yes.

Ask again later.

Better not to tell you now.

Cannot predict now. 

Concentrate and ask again.

Don't count on it. 

It is certain.

It is decidedly so.

Most likely.

My reply is no. 

My sources say no.

Outlook not so good.

Outlook good. 

Reply hazy, try again.

Signs point to yes.

Very doubtful.

Without a doubt.

Yes.

Yes - definitely. 

You may rely on it.

You do not need to have so many responses. Instead, choose at least 5 (or make up your own!). 

Note for me: (not completed success this task yet, I'll do it more later)